code-samples/Python/autogen-planning-agent/requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
autogen-core>=0.2.0
autogen-ext>=0.2.0
azure-core>=1.29.0
azure-ai-inference
aiohttp